# api/openapi.yaml
openapi: 3.1.0
info:
  title: model-manager API
  version: 1.0.0
  description: |
    Backend-agnostic model management for the Giant Swarm Agent Platform.

    One API implemented by serving backends (`ollama` for laptop/agentlab
    installs, `kserve` for GPU installs, `lemonade` for hosts running Lemonade
    Server — FastFlowLM on AMD Ryzen AI NPUs, llama.cpp on GPU and CPU —
    and `lmstudio` for hosts running LM Studio, which serves no delete). One
    model-manager runs one or several of them at once (`--backends`, chart
    value `backends`). Clients MUST call `GET /api/v1/backends` first and
    render only what each backend's capability flags allow: a false flag
    means the matching operation answers `501 unsupported` on that backend.
    `GET /api/v1/backend` stays for clients of one backend: it describes the
    default backend (the first configured) and names the others in
    `backends`.

    Every model, loaded entry, job, node, preset, fit result and ModelConfig
    reference carries `backend`. Reads take an optional `?backend=` filter and
    otherwise aggregate every backend — a backend that fails to answer is
    reported in the response's `errors` while the others' items are returned
    (`502` only when every backend failed). Mutations take an optional
    `backend` in the body (or query, on `DELETE /api/v1/models/{name}`); a
    request without it is resolved to the one backend holding the reference —
...

...
    in a JSON body instead.

...

...
servers:
...
paths:
...
  /api/v1/backend:
    get:
      tags: [backend]
      summary: One backend — the named one, else the default — plus the names of every configured backend
      description: |
        `endpoint` is the backend as model-manager dials it; `agentEndpoint`
        the backend as agent pods dial it (what ModelConfigs point at), when
        the backend has one. Render controls per capability flag. `loading`
        says how the backend brings models into memory and lets them go —
        use it to word a not-loaded model ("loads on first request" versus
        "not serving"), never the backend name. `backends` names every
        configured backend; a client that sees more than one switches to
        `GET /api/v1/backends`.
      operationId: getBackend
      x-mcp-tool: get_backend
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
      responses:
        "200":
          description: backend descriptor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Backend" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
...
  /api/v1/models:
    get:
      tags: [models]
      summary: Downloaded models (inventory) of one or every backend, enriched with loaded state and ModelConfig
      operationId: listModels
      x-mcp-tool: list_models
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
      responses:
        "200":
          description: inventory; `errors` names the backends that could not be read (their models are missing)
          content:
            application/json:
              schema:
                type: object
                required: [models]
                properties:
                  models:
                    type: array
                    items: { $ref: "#/components/schemas/Model" }
                  errors: { $ref: "#/components/schemas/BackendErrors" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/models/{name}:
    parameters:
      - $ref: "#/components/parameters/ModelName"
      - $ref: "#/components/parameters/BackendQuery"
    get:
      tags: [models]
      summary: One downloaded model
      operationId: getModel
      x-mcp-tool: get_model
      responses:
        "200":
          description: model
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Model" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Ambiguous" }
        "502": { $ref: "#/components/responses/BackendError" }
    delete:
      tags: [models]
      summary: Delete a downloaded model (and by default its ModelConfig)
      operationId: deleteModel
      x-mcp-tool: delete_model
      parameters:
        - name: unwire
          in: query
          description: Also remove the ModelConfig. Default true.
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  backend: { type: string, description: the backend the model was deleted from }
                  model: { type: string }
                  deleted: { type: boolean }
                  unwired: { type: boolean }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Ambiguous" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/loaded:
    get:
      tags: [models]
      summary: Models currently loaded / serving, on one or every backend that lists them
      operationId: listLoaded
      x-mcp-tool: list_loaded_models
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
      responses:
        "200":
          description: loaded models; `errors` names the backends that could not be read
          content:
            application/json:
              schema:
                type: object
                required: [loaded]
                properties:
                  loaded:
                    type: array
                    items: { $ref: "#/components/schemas/LoadedModel" }
                  errors: { $ref: "#/components/schemas/BackendErrors" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/models/pull:
    post:
      tags: [models, jobs]
      summary: Start importing a model; returns a job immediately
      description: |
        Starts a background import and returns `202` with the job. If an import
        of the same reference is already pending/running, that job is returned
        with `created: false`. Poll `GET /api/v1/jobs/{id}` for progress. On
        success the model is wired into kagent (job `result` carries the
        ModelConfig reference) unless `wire` is false or the server's autoWire
        is off. On kserve the job echoes `preset` and `node`: the values the
        request named right away, and the node model-manager picked itself
        once the fit check ran. The pull runs on `backend`, else on the
        default backend (the first configured): a pull names a model that
        does not exist yet, so nothing else can pick the backend.
      operationId: pullModel
      x-mcp-tool: pull_model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model]
              properties:
                backend: { $ref: "#/components/schemas/BackendName" }
                model:
                  type: string
                  description: "Model reference (ollama registry tag or hf.co/... GGUF ref; lemonade catalog model name such as Qwen3-0.6B-GGUF — `lemonade list`; lmstudio hub reference such as ibm/granite-4-micro — `lms ls`; kserve Hugging Face repository owner/name[:revision])"
                  example: smollm2:135m
                wire:
                  type: boolean
                  description: Create a ModelConfig on completion (ollama, lemonade, lmstudio; default = server autoWire; kserve refuses true — models are wired when served)
                preset:
                  type: string
                  description: kserve — the serving preset the download is for; its InferenceService mounts the resulting cache directory. Default = the single preset serving the model.
                node:
                  type: string
                  description: "kserve — node whose cache receives the download; must be eligible (GET /api/v1/nodes), else the job fails with does_not_fit and the eligibility reason. Default = the cache node, else the eligible node with the largest budget."
      responses:
        "202":
          description: job accepted (or joined)
          content:
            application/json:
              schema:
                type: object
                required: [job, created]
                properties:
                  job: { $ref: "#/components/schemas/Job" }
                  created: { type: boolean }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "501": { $ref: "#/components/responses/Unsupported" }
  /api/v1/models/load:
    post:
      tags: [models]
      summary: Load a model into memory / start serving it
      description: |
        ollama: a keep-alive load that pre-warms the model — Ollama re-arms
        the eviction timer on every later request with that request's own
        keep-alive (or the server's `OLLAMA_KEEP_ALIVE` when it sends none),
        so the `keepAlive` given here holds only until the next inference
        request (`loading.keepAliveScope: request`). The response is the
        loaded model with `running.expiresAt`, the deadline Ollama reports
        right after the load. kserve: fit-checks the model, composes an
        InferenceService from the preset (the only way to serve on kserve) and
        creates it; the response has `loaded=true` with `running.status`
        `Pending` until the predictor is ready, and a `load` job (see
        `GET /api/v1/jobs`) follows readiness and wires the model into kagent.
        Loading an already-served preset again is a no-op. lemonade: `POST
        /api/v1/load` on the Lemonade Server, which starts the model's
        backend process (FastFlowLM on the NPU, llama.cpp, ...) and answers
        once it is ready — no keep-alive: the model stays until unloaded or
        until another model of its type needs the slot; `keepAlive: "-1"`
        pins it against that displacement, other values are ignored. The
        response is the loaded model with `running.device` and
        `running.pinned`. lmstudio: `POST /api/v1/models/load`, which reads
        the weights and answers with the resident instance; the load persists
        until unloaded (LM Studio exempts an explicit load from its idle TTL
        and from Auto-Evict) and `keepAlive` has nothing to map onto, since
        the endpoint takes no TTL. Loading an already-resident model is a
        no-op: LM Studio loads an instance, not a model, so a repeat would
        pin a second copy.
      operationId: loadModel
      x-mcp-tool: load_model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backend: { $ref: "#/components/schemas/BackendName" }
                model: { type: string, description: model reference; required unless preset is given }
                keepAlive:
                  type: string
                  description: |
                    ollama keep-alive after this load (duration such as 10m; -1 = forever).
                    Default = model-manager's `--default-keep-alive`, reported as
                    `loading.keepAliveDefault`. Later inference requests replace it
                    with their own keep-alive. lemonade: only `-1` has a meaning —
                    it pins the model against slot eviction; every other value is
                    ignored (no idle timer). lmstudio: ignored entirely — its
                    load endpoint accepts no TTL, and an explicitly loaded
                    model has no idle timer to set.
                preset:
                  type: string
                  description: kserve — serving preset to compose the InferenceService from. Default = the single preset serving the model.
                node:
                  type: string
                  description: "kserve — pin the predictor to this node; must be eligible (GET /api/v1/nodes), else 412 does_not_fit with the eligibility reason and no InferenceService is created."
      responses:
        "200":
          description: the loaded model — `loaded=true`, `running` as the backend reports it right after the load (ollama with `expiresAt`; lemonade with `device` and `pinned`; kserve with `status` Pending)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Model" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: an InferenceService of that name exists for another model or is not managed by model-manager; or the reference exists on several backends and none was named
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "412": { $ref: "#/components/responses/DoesNotFit" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/models/fit-check:
    post:
      tags: [serving]
      summary: Check whether a model fits a node before downloading or serving it (kserve)
      description: |
        Resolves the weight size from the Hugging Face Hub
        (`model.safetensors.index.json`, else the file tree, else the preset's
        `weightsGiB`), adds the serving overhead (the preset's `overheadGiB`,
        default 30 GiB) and compares the sum with the memory budget of the
        target node (GPU memory from the node labels when present, else
        allocatable memory; the node annotation
        `model-manager.giantswarm.io/memory-budget-gib` overrides it —
        `budgetSource` says which applied). `fits=false` is a 200 with the
        explanation in `reason`; pull and load refuse such models with
        `412 does_not_fit`. A `node` that is not a serving target
        (`eligible: false` on `GET /api/v1/nodes`, or a CPU-only node that is
        not listed there) answers `fits=false` with the eligibility reason
        and no `node`; pull and load refuse it the same way, before any Job
        or InferenceService is created. Without a `node` only eligible nodes
        are considered.
      operationId: fitCheck
      x-mcp-tool: check_fit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backend: { $ref: "#/components/schemas/BackendName", description: "the backend to check on; required only when several configured backends offer fit checks (else 409)" }
                model: { type: string, description: Hugging Face repository owner/name; required unless preset is given }
                preset: { type: string }
                node: { type: string, description: "node to check against — must be eligible (GET /api/v1/nodes); default = the eligible node with the most free budget, cache nodes first" }
      responses:
        "200":
          description: fit result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FitResult" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Ambiguous" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/presets:
    get:
      tags: [serving]
      summary: Curated serving presets (kserve), of one or every backend that offers them
      description: |
        The published `ServingPreset` ConfigMaps of the platform
        (`agent-platform.giantswarm.io/serving-preset=true`), resolved to what a
        client needs to pick one: model, runtime, GPUs, weights + overhead
        requirements, arguments.
      operationId: listPresets
      x-mcp-tool: list_presets
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
      responses:
        "200":
          description: presets
          content:
            application/json:
              schema:
                type: object
                required: [presets]
                properties:
                  presets:
                    type: array
                    items: { $ref: "#/components/schemas/Preset" }
                  errors: { $ref: "#/components/schemas/BackendErrors" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/search:
    get:
      tags: [serving]
      summary: Search the model hub (kserve — Hugging Face Hub) of one or every backend that has one
      operationId: searchModels
      x-mcp-tool: search_models
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 50 }
      responses:
        "200":
          description: search results, most downloaded first
          content:
            application/json:
              schema:
                type: object
                required: [results]
                properties:
                  query: { type: string }
                  results:
                    type: array
                    items: { $ref: "#/components/schemas/SearchResult" }
                  errors: { $ref: "#/components/schemas/BackendErrors" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/nodes:
    get:
      tags: [serving]
      summary: Node memory budgets, reservations and download caches
      description: |
        kserve: the accelerator nodes only — nodes that advertise the
        configured GPU resource (`gpuResourceName` of the discovery ConfigMap,
        default `nvidia.com/gpu`; capacity or allocatable > 0) or carry a
        gpu-feature-discovery label (`nvidia.com/gpu.present`, `.count`,
        `.product`); CPU-only nodes are not serving capacity and are left
        out. Each node comes with its memory budget for fit checks
        (`budgetSource`: `gpu-labels` from the NVIDIA feature-discovery labels,
        `allocatable` memory, or `annotation` when the node carries
        `model-manager.giantswarm.io/memory-budget-gib: "<GiB>"` — the
        per-node override for unified-memory nodes), what running models
        reserve, the download cache the node holds, and `eligible`: whether a
        model can be served there right now — the node is ready, inside the
        serving `nodeSelector` of the discovery ConfigMap, and can mount the
        cache claim whenever predictors mount it (cache enabled with
        `cache.redirectPolicy` on) and the claim is pinned to nodes (a static
        local PersistentVolume). `eligibilityReason` explains a false value,
        e.g. `outside the serving node selector
        (kubernetes.io/hostname=spark-8723); cache claim hf-cache is pinned to
        spark-8723` or `not ready`; pull, load and fit-check refuse such a
        node with that reason. `message` reports an ignored, unparsable
        annotation.

        ollama: one node, the proxied host (`name` = the host agents dial,
        e.g. `172.21.0.1`). Ollama's API has no capacity endpoint, so
        `budgetBytes` (= `allocatableMemoryBytes`) is `MemTotal` of
        `/proc/meminfo` as the model-manager pod sees it (`budgetSource:
        host-meminfo`; the machine's RAM when pod and Ollama share a kernel,
        the VM's under Docker Desktop or another VM-backed runtime) unless the
        operator set the chart value `ollama.memoryBudgetGiB`
        (`--ollama-memory-budget-gib`): then `budgetBytes` is that figure with
        `budgetSource: override` while `allocatableMemoryBytes` stays the
        pod's `MemTotal` — the counterpart of the kserve node annotation; a
        value that is not a positive number of GiB is ignored and named in
        `message`. `reservedBytes` is the sum of `size` over Ollama's `/api/ps`,
        `accelerated` says whether any loaded model has memory on an
        accelerator (`size_vram` > 0). `gpuCount` is 0 — accelerators are not
        observable through Ollama's API; the per-model share is
        `running.vramBytes` on `/models` and `/loaded` — and there is no
        `cache`. `message` explains the figures.

        lemonade: one node, the proxied host (`name` = the host agents dial),
        described by Lemonade's own `GET /api/v1/system-info`: `budgetBytes`
        (= `allocatableMemoryBytes`) is the host memory Lemonade reports
        (`budgetSource: system-info`), `gpuCount` and `gpuProduct` are the
        accelerators Lemonade found available — the NPU (named) and the GPUs
        — `reservedBytes` is the sum of the catalog sizes of the loaded models
        (Lemonade reports no per-model memory), and `cache` is Lemonade's
        model store (`mountPath`, the downloaded models and their sizes,
        `inventory: models`). Always eligible.
      operationId: listNodes
      x-mcp-tool: list_nodes
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
      responses:
        "200":
          description: nodes of one or every backend that reports them; `errors` names the backends that could not be read
          content:
            application/json:
              schema:
                type: object
                required: [nodes]
                properties:
                  nodes:
                    type: array
                    items: { $ref: "#/components/schemas/NodeInfo" }
                  errors: { $ref: "#/components/schemas/BackendErrors" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
  /api/v1/models/unload:
    post:
      tags: [models]
      summary: Unload a model from memory / stop serving it
      operationId: unloadModel
      x-mcp-tool: unload_model
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModelRequest" }
      responses:
        "200":
          description: unloaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  backend: { type: string, description: the backend the model was unloaded from }
                  model: { type: string }
                  loaded: { type: boolean, const: false }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Ambiguous" }
        "501": { $ref: "#/components/responses/Unsupported" }
        "502": { $ref: "#/components/responses/BackendError" }
...
  /api/v1/jobs:
    get:
      tags: [jobs]
      summary: List jobs, newest first, of one or every backend
      operationId: listJobs
      x-mcp-tool: list_jobs
      parameters:
        - $ref: "#/components/parameters/BackendQuery"
      responses:
        "200":
          description: jobs
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
...
components:
  parameters:
...
    BackendQuery:
      name: backend
      in: query
      required: false
      description: "Restrict the operation to one configured backend (`GET /api/v1/backends` lists them). Reads without it aggregate every backend; a reference without it is resolved to the one backend holding it. An unknown or unconfigured name answers 400."
      schema: { $ref: "#/components/schemas/BackendName" }
...